added multibyte address support

+++ app/src/crypto_helper.h
@@ -31,7 +31,7 @@ extern "C" {
 #define SS58_ADDRESS_MAX_LEN 60u
 
 uint8_t crypto_SS58EncodePubkey(uint8_t *buffer, uint16_t buffer_len,
-                                uint8_t addressType, const uint8_t *pubkey);
+                                uint16_t addressType, const uint8_t *pubkey);
 
 #ifdef __cplusplus
 }
